chore: organize and refactor database impl
Setup the server code and refactored the database to use a pool. Moved
the database to a ServerDependencies struct and wrapped in an `Arc` In
follow up commits we should be able to use this pool to get a connection
inside the auth service and write to the database there.

refactored code into modules and formatted with `cargo +nightly fmt`