Moss street/7/server configurations (#9)
* feat: setup server proto stuff removed

can't get proto stuff to build in but was able to try some server
changes by bringing it in manually. Removed the proto files so this
won't build. Next commit will focus on bringing in the repo and getting
it to build

* chore: organize and refactor database impl

Setup the server code and refactored the database to use a pool. Moved
the database to a ServerDependencies struct and wrapped in an `Arc` In
follow up commits we should be able to use this pool to get a connection
inside the auth service and write to the database there.

refactored code into modules and formatted with `cargo +nightly fmt`

* fix: Install protobuf compiler for actions script

---------

Co-authored-by: Justyn Lantz <jtlantz@amazon.com>