feat: add plumbing for wallets (#28)
wallets are going to be constructs that can contain any amount of any stock. In this case it will allow us to use the stock object (probably later to be renmaed) and contain our actual currency as well as stock amounts. E.G. we will create a USD stock, and a APPL stock, then trade on those pairs and swap USD for APPL.

Co-authored-by: Justyn lantz <jtlantz@github.com>