feat: start implementing market order book
The current implementation for this is pretty bad but this is a good
foundation to start building the order book processor. There is a vision
for the future implementation that uses channels between the api handler
and the market handler but that will take a bit more planning. this
commit is not meant to merged to main.